Store territory in config
When using LanguageSelectionPanel.set_config_values(), store the territory as part of the locale name
if it has one, separated by _. For example, store en_US for US english. This more completely matches the display name.
We do this because two locales for the same language (en_US and en_GB) have different time formats, and just using the base name causes wx to pick the wrong one, at least for English.